layout/base/tests/chrome/test_will_change.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tests for MozAfterPaint</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/paint_listener.js"></script>
  <style>
    #checkOpacityRepaint {
      will-change: opacity;
    }
    #checkTransformRepaint {
      will-change: transform;
    }
    #checkTranslateRepaint {
      will-change: translate;
    }
    div {
      width: 100px;
      height: 100px;
      background: radial-gradient(ellipse at center, #87e0fd 0%,#53cbf1 40%,#05abe0 100%);
    }
  </style>
</head>
<body>
  <div id="checkRepaint">
    Check repaint without will-change
  </div>
  <div id="checkOpacityRepaint">
    Check repaint with will-change
  </div>
  <div id="checkTransformRepaint">
    Check repaint with will-change
  </div>
  <div id="checkTranslateRepaint">
    Check repaint with will-change
  </div>
</body>
<script>

var utils = window.windowUtils;

function waitForPaints() {
  return new Promise(function(resolve, reject) {
    waitForAllPaintsFlushed(resolve);
  });
}

add_task(async () => {
  var element = document.getElementById("checkRepaint");

  await waitForPaints();

  utils.checkAndClearPaintedState(element);
  element.style.opacity = "0.5";

  await waitForPaints();

  var painted = utils.checkAndClearPaintedState(element);
  // *** We check that this repaints because the test is relying
  //     on this property. If this is broken then this test wont
  //     be reliable check for will-change.
  is(painted, true, "element should have been painted");
});

add_task(async () => {
  var element = document.getElementById("checkOpacityRepaint");

  await waitForPaints();

  utils.checkAndClearPaintedState(element);
  element.style.opacity = "0.5";

  await waitForPaints();

  var painted = utils.checkAndClearPaintedState(element);
  // BasicLayers' heuristics are so that even with will-change:opacity,
  // we can still have repaints.
  if (utils.layerManagerType != "Basic") {
    is(painted, false, "will-change checkOpacityRepaint element should not have been painted");
  }
});

add_task(async () => {
  var element = document.getElementById("checkTransformRepaint");

  await waitForPaints();

  utils.checkAndClearPaintedState(element);
  element.style.transform = "translateY(-5px)";

  await waitForPaints();

  var painted = utils.checkAndClearPaintedState(element);
  // BasicLayers' heuristics are so that even with will-change:transform,
  // we can still have repaints.
  if (utils.layerManagerType != "Basic") {
    is(painted, false, "will-change checkTransformRepaint element should not have been painted");
  }
});

add_task(async () => {
  var element = document.getElementById("checkTranslateRepaint");

  await waitForPaints();

  utils.checkAndClearPaintedState(element);
  element.style.translate = "5px";

  await waitForPaints();

  var painted = utils.checkAndClearPaintedState(element);
  // BasicLayers' heuristics are so that even with will-change:translate,
  // we can still have repaints.
  if (utils.layerManagerType != "Basic") {
    is(painted, false, "will-change checkTranslateRepaint element should not have been painted");
  }
});

</script>
</html>
